updated top populated cities in the region (fixed data type)

+++ src/main/java/com/napier/coursework/City.java
@@ -22,7 +22,7 @@ public class City {
         District = district;
     }
 
-    public void setPopulation(double population) {
+    public void setPopulation(int population) {
         Population = population;
     }
